Updated files to 2.2.0 - NVIDIA GPU issues fixed
add 3 donation types to myself and removed orig creators addresses.
Again i do not want to take this away from them - if you wish to donate extra to those home take your 2% donation already, feel free to search their Git and donate. Otherwise - i have taken my time to compile this for you all, i do not promote this, i do not wish for it to be promoted.

give this to your friends and friends of friends and lets keep the community small but together - Cryto needs to take off giving everyone everything for free as this world goes, is Terrible- and yes many of months ago i already made a small donation to the original creators of this software.